Work on interpreter

+++ source/debug.hpp
@@ -5,6 +5,6 @@ namespace Lux {
     class Chunk;
 
     void disassembleChunk(const Chunk& chunk, const char* name);
-    int disassembleInstruction(const Chunk& chunk, int offset);
+    size_t disassembleInstruction(const Chunk& chunk, size_t offset);
 
 } // namespace Lux
